+++ slips_files/core/database/_profile_flow.py
@@ -11,8 +11,6 @@ class ProfilingFlowsDatabase(object):
         # The name is used to print in the outputprocess
         self.name = 'DB'
         self.separator = '_'
-        # flag to know which flow is the start of the pcap/file
-        self.first_flow = True
         self.seen_MACs = {}
 
 
@@ -808,6 +806,8 @@ class ProfilingFlowsDatabase(object):
         spkts='',
         sbytes='',
         appproto='',
+        smac='',
+        dmac='',
         uid='',
         label='',
         flow_type='',
@@ -815,6 +815,7 @@ class ProfilingFlowsDatabase(object):
         """
         Function to add a flow by interpreting the data. The flow is added to the correct TW for this profile.
         The profileid is the main profile that this flow is related too.
+        : param new_profile_added : is set to True for everytime we see a new srcaddr
         """
         summaryState = self.getFinalStateFromFlags(state, pkts)
         data = {
@@ -832,11 +833,12 @@ class ProfilingFlowsDatabase(object):
             'spkts': spkts,
             'sbytes': sbytes,
             'appproto': appproto,
+            'smac': smac,
+            'dmac': dmac,
             'label': label,
             'flow_type': flow_type,
             'module_labels': {},
         }
-        # when adding a flow, there are still no labels ftom other modules, so the values is empty dictionary
 
         # Convert to json string
         data = json.dumps(data)
@@ -854,24 +856,43 @@ class ProfilingFlowsDatabase(object):
         # Store the label in our uniq set, and increment it by 1
         if label:
             self.r.zincrby('labels', 1, label)
+
         # We can publish the flow directly without asking for it, but
         # its good to maintain the format given by the get_flow() function.
         flow = self.get_flow(profileid, twid, uid)
+
         # Get the dictionary and convert to json string
         flow = json.dumps(flow)
         # Prepare the data to publish.
-        to_send = {}
-        to_send['profileid'] = profileid
-        to_send['twid'] = twid
-        to_send['flow'] = flow
-        to_send['stime'] = stime
+        to_send = {
+            'profileid': profileid,
+            'twid': twid,
+            'flow': flow,
+            'stime': stime,
+        }
         to_send = json.dumps(to_send)
+
         # set the pcap/file stime in the analysis key
         if self.first_flow:
-            self.first_flow = False
             self.set_input_metadata({'file_start': stime})
+            self.first_flow = False
+
+        # set the local network used in the db
+        if not self.is_localnet_set:
+            if (
+                    validators.ipv4(saddr)
+                    and ipaddress.ip_address(saddr).is_private
+            ):
+                # get the local network of this saddr
+                if network_range := utils.get_cidr_of_ip(saddr):
+                    self.r.set("local_network", network_range)
+                    self.is_localnet_set = True
+
         self.publish('new_flow', to_send)
         return True
+
+    def get_local_network(self):
+         return self.r.get("local_network")
 
     def get_label_count(self, label):
         """